+++ NEWS.md
@@ -1,5 +1,13 @@
 # odbc (development version)
 
+* Added a function `odbcListConfig()` to help locate configuration files on
+  macOS and Linux (@simonpcouch, #565).
+
+* Use correct parent class for Oracle (#685).
+
+
+# odbc 1.4.0
+
 ## Major changes
 
 * New `odbc::databricks()` makes it easier to connect to Databricks,